Extract Counts Matrix from Molecule Experiment based on Assay Type

This function retrieves a counts matrix from a Molecule Experiment object based on the given assay type.

CountsMatrix(me, assayName, nCores = 1, ...)

Arguments

me

A Molecule Experiment object.

assayName

A character string indicating the assay type. Supported values include "sub-sector", "sub-concentric", "sub-combo", "super-concentric", and "super-combo".

nCores

Number of cores for parallel processing (default 1)

...

arguments passing to MoleculeExperiment::countMolecules

Value

A counts matrix corresponding to the specified assay type.

Examples

if (FALSE) {
# Assuming `data_obj` is your Molecule Experiment object
cm = CountsMatrix(data_obj, assayName = "sub-sector")
}