Create a SummarizedExperiment Object from a List of Data Frames

This function takes a list of data frames and a Molecule Experiment object and constructs a SummarizedExperiment object with the specified data structure.

Usage

EntropySummarizedExperiment(
  df_list,
  me,
  includeCounts = FALSE,
  concatenateFeatures = FALSE,
  nCores = 1
)

Arguments

df_list

A list of data frames, each containing assay data.

me

A Molecule Experiment object.

includeCounts

logical (default FALSE) whether to include gene counts as features

concatenateFeatures

logical whether to concatenate all the features into a single assay (default FALSE). If FALSE the output SE object has multiple assays

nCores

Number of cores (default 1)

Value

A SummarizedExperiment object.

Examples

data(example_me)
me <- loadBoundaries(me)
ent <- EntropyMatrix(me, c("subsector", "subconcentric", "supersector", "superconcentric"), nCores = 1)
se <- EntropySummarizedExperiment(ent, me)
se
#> class: SummarizedExperiment 
#> dim: 178 14 
#> metadata(0):
#> assays(4): subsector subconcentric supersector superconcentric
#> rownames(178): 2010300C02Rik Acsbg1 ... Zfp536 Zfpm2
#> rowData names(0):
#> colnames(14): sample1.67500 sample1.67512 ... sample2.65070
#>   sample2.65071
#> colData names(5): Cell Sample_id x_central y_central boundaries