Generate Feature Data Based on Assay Type Combining Various Polygon Assays

This comprehensive function processes the given data (me) and returns feature data based on multiple assay types. The supported assays encapsulated within this function include:

Sub-sector polygons

Sub-concentric polygons

Sub-combo polygons

Super-concentric polygons

Super-combo polygons

These assays represent various spatial subdivisions or transformations, each addressing unique feature characteristics. The results from each assay are combined and returned as a list of dataframes.

Usage

GenerateFeatureData(
  me,
  featureTypes = c("subsector", "subconcentric", "supersector", "superconcentric"),
  k = 5
)

Arguments

me

The input data object. This object contains spatial information, typically regarding segments or regions of interest.

featureTypes

Character vector containing the feature types to calculate.

k

A numeric value indicating the scaling factor for concentric polygons. Defaults to 5.

Value

A list containing dataframes for each assay type:

subsector: Feature data for sub-sector polygons.

subconcentric: Feature data for sub-concentric polygons.

supersector: Feature data for super-sector polygons.

superconcentric: Feature data for super-concentric polygons.