Calculate SpatialFeatures

This function takes a MoleculeExperiment (ME) object as input and calculates new molecule-based spatial features according to the following segmentations:

Sub-sector polygons

Sub-concentric polygons

Super-sector polygons

Super-concentric polygons

The function has 3 key steps, 1. adds new boundaries to the ME object according to the subcellular and supercellular segmentations, 2. calculates entropy for each of the cell and featureType combinations, and 3. combines the entropy values into a SummarizedExperiment object.

spatialFeatures(
  me,
  featureTypes = c("sub_sector", "sub_concentric", "super_sector", "super_concentric"),
  k = 5,
  nCores = 1,
  includeCounts = FALSE,
  ...
)

Arguments

me

A MoleculeExperiment (ME) object

featureTypes

a character vector listing the types of featureTypes to include

k

A numeric value indicating the number of polygons to calculate entropy (default 5)

nCores

number of cores for parallel processing (default 1)

includeCounts

logical (default FALSE) whether to include gene counts as features

...

arguments passed to loadBoundaries and EntropyMatrix

Value

A SummarizedExperiment object containing a spatialFeatures assay and cell-level colData

Examples

data(example_me)
se <- spatialFeatures(me)
se
#> class: SummarizedExperiment 
#> dim: 712 14 
#> metadata(0):
#> assays(1): spatialFeatures
#> rownames(712): sub_sector_2010300C02Rik sub_sector_Acsbg1 ...
#>   super_concentric_Zfp536 super_concentric_Zfpm2
#> rowData names(2): FeatureCategory FeatureGene
#> colnames(14): sample1.67500 sample1.67512 ... sample2.65070
#>   sample2.65071
#> colData names(5): Cell Sample_id x_central y_central boundaries